md-graph/requirements.txt
backcall==0.2.0
decorator==4.4.2
ipython==7.31.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.3
jsonpickle==1.4.2
MarkupSafe==1.1.1
networkx==2.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.4
pyvis==0.1.8.2
PyYAML==5.4.1
traitlets==5.4.0
wcwidth==0.2.5
configargparse==1.5.3